bc_obps/pyproject.toml
[tool.poetry]
name = "registration"
version = "0.1.0"
description = ""
authors = ["Andrea Williams <andrea.williams@gov.bc.ca>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.7"
python-dotenv = "^1.0.0"
django-ninja = "^1.1.0"
psycopg = "^3.1.10"
psycopg-binary = "^3.1.10"
django-phonenumber-field = "^7.1.0"
django-phonenumbers = "^1.0.1"
django-storages = {extras = ["google"], version = "^1.14"}
whitenoise = "^6.5.0"
pre-commit = "^3.4.0"
django-cors-headers = "4.2.0"
django-localflavor = "^4.0"
pytz = "^2023.3.post1"
django-simple-history = "^3.4"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^5.2"
pylint = "^2.17.5"
pytest-django = "^4.5.2"
pytest-watch = "^4.2.0"
pytest-testmon = "^2.0.13"
pytest-picked = "^0.5.0"
model-bakery = "^1.16.0"
django-extensions = "^3.2.3"
ipdb = "^0.13.13"
pytest-cov = "^4.1.0"
django-diagram = "^0.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
exclude = '''
    /(
        \.git          # exclude the ".git" directory
        | \.hg           # exclude the ".hg" directory
        | \.mypy_cache   # exclude the ".mypy_cache" directory
        | \.tox          # exclude the "tox" virtualenv directory
        | \.venv         # exclude the "venv" virtualenv directory
        | _build         # exclude the "_build" directory
        | buck-out       # exclude the "buck-out" directory
        | build          # exclude the "build" directory
        | dist           # exclude the "dist" directory
        | node_modules   # exclude the "node_modules" directory
    )/
'''
